+++ mayan/apps/documents/api_views/mixins.py
@@ -45,8 +45,15 @@ class ParentObjectDocumentTypeAPIViewMixin:
 
 
 class ParentObjectDocumentVersionAPIViewMixin(ParentObjectDocumentAPIViewMixin):
-    def get_document_version(self):
+    def get_document_version(self, permission=None):
+        queryset=self.get_document().versions.all()
+
+        if permission:
+            queryset = AccessControlList.objects.restrict_queryset(
+                permission=permission, queryset=queryset,
+                user=self.request.user
+            )
+
         return get_object_or_404(
-            queryset=self.get_document().versions.all(),
-            pk=self.kwargs['document_version_id']
+            queryset=queryset, pk=self.kwargs['document_version_id']
         )