+++ mayan/apps/documents/tests/mixins/document_type_mixins.py
@@ -28,7 +28,7 @@ class DocumentQuickLabelViewTestMixin:
 
 class DocumentTypeAPIViewTestMixin:
     def _request_test_document_type_create_api_view(self):
-        pk_list = list(DocumentType.objects.values('pk'))
+        pk_list = list(DocumentType.objects.values_list('pk', flat=True))
 
         response = self.post(
             viewname='rest_api:documenttype-list', data={